requirements.txt
groq==0.9.0
python-dotenv==1.0.1
transformers==4.46.3
torch==2.4.1
accelerate==1.0.1

